backoffice and public restructured
