+++ src/app/features/backoffice/backoffice.component.html
@@ -1,0 +1,2 @@
+<router-outlet></router-outlet> 
+<h1>hola</h1>
