+++ src/app/features/backoffice/backoffice.component.ts
@@ -1,0 +1,14 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'alk-backoffice',
+  templateUrl: './backoffice.component.html',
+})
+export class BackofficeComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}
