+++ src/app/features/public/pages/home/home.component.html
@@ -1,5 +1,5 @@
 
-<alk-header></alk-header>
+<!-- <alk-header></alk-header> -->
 <div class="container">
     <h1> Bienvenidos</h1>
     <h2> {{description }}</h2>
@@ -13,4 +13,4 @@
     </div>
 </div>
 
-<alk-footer></alk-footer>
+<!-- <alk-footer></alk-footer> -->
