+++ src/app/features/public/pages/register-form/register-form.component.scss
