+++ src/app/features/public/public.component.html
@@ -1,0 +1,4 @@
+<alk-header></alk-header>
+<router-outlet></router-outlet>
+
+
